control-plane/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
grpcio==1.78.0
grpcio-tools==1.78.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
alembic==1.18.4
redis==7.4.0
json-logic-qubit==0.9.1
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==1.3.0